Re #16 Fix generation of massive logfile when EDMC is restarted.
Fixes the issue that when the game and overlay were running, restarting EDMC would cause the
overlay server to emit never-ending log messages recording that EDMC had quit. This could continue until the disk was full or the overlay was stopped.

Special thanks to CorporalLeak and Austeregrim for helping to debug
this.